+++ plugin/trino-iceberg/src/test/java/io/trino/plugin/iceberg/TestIcebergMaterializedView.java
@@ -26,7 +26,6 @@ import java.util.Optional;
 
 import static io.trino.plugin.base.util.Closables.closeAllSuppress;
 import static io.trino.plugin.iceberg.IcebergQueryRunner.ICEBERG_CATALOG;
-import static io.trino.testing.TestingNames.randomNameSuffix;
 import static org.apache.iceberg.BaseMetastoreTableOperations.METADATA_LOCATION_PROP;
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -85,170 +84,6 @@ public class TestIcebergMaterializedView
     {
         Table table = metastore.getTable("tpch", materializedViewName).orElseThrow();
         return table.getParameters().get(METADATA_LOCATION_PROP);
-    }
-
-    @Test
-    public void testIncrementalRefresh()
-    {
-        String sourceTableName = "source_table" + randomNameSuffix();
-        String materializedViewName = "test_materialized_view_" + randomNameSuffix();
-
-        Session defaultSession = getSession();
-        Session incrementalRefreshDisabled = Session.builder(getSession())
-                .setCatalogSessionProperty("iceberg", "incremental_refresh_enabled", "false")
-                .build();
-
-        String matViewDef = "SELECT a, b FROM %s WHERE a < 3 OR a > 5".formatted(sourceTableName);
-
-        // create source table and two identical MVs
-        assertUpdate("CREATE TABLE %s (a int, b varchar)".formatted(sourceTableName));
-        assertUpdate("INSERT INTO %s VALUES (1, 'abc'), (2, 'def')".formatted(sourceTableName), 2);
-        assertUpdate("CREATE MATERIALIZED VIEW %s_1 AS %s".formatted(materializedViewName, matViewDef));
-        assertUpdate("CREATE MATERIALIZED VIEW %s_2 AS %s".formatted(materializedViewName, matViewDef));
-
-        // execute first refresh: afterwards both MVs will contain: (1, 'abc'), (2, 'def')
-        assertUpdate("REFRESH MATERIALIZED VIEW %s_1".formatted(materializedViewName), 2);
-        assertUpdate("REFRESH MATERIALIZED VIEW %s_2".formatted(materializedViewName), 2);
-
-        // add some new rows to source
-        assertUpdate("INSERT INTO %s VALUES (3, 'ghi'), (4, 'jkl'), (5, 'mno'), (6, 'pqr')".formatted(sourceTableName), 4);
-
-        // will do incremental refresh, and only add: (6, 'pqr')
-        assertUpdate(defaultSession, "REFRESH MATERIALIZED VIEW %s_1".formatted(materializedViewName), 1);
-        // will do full refresh, and (re)add: (1, 'abc'), (2, 'def'), (6, 'pqr')
-        assertUpdate(incrementalRefreshDisabled, "REFRESH MATERIALIZED VIEW %s_2".formatted(materializedViewName), 3);
-
-        // verify that view contents are the same
-        assertThat(query("TABLE %s_1".formatted(materializedViewName))).matches("VALUES (1, VARCHAR 'abc'), (2, VARCHAR 'def'), (6, VARCHAR 'pqr')");
-        assertThat(query("TABLE %s_2".formatted(materializedViewName))).matches("VALUES (1, VARCHAR 'abc'), (2, VARCHAR 'def'), (6, VARCHAR 'pqr')");
-
-        // cleanup
-        assertUpdate("DROP MATERIALIZED VIEW %s_1".formatted(materializedViewName));
-        assertUpdate("DROP MATERIALIZED VIEW %s_2".formatted(materializedViewName));
-        assertUpdate("DROP TABLE %s".formatted(sourceTableName));
-    }
-
-    @Test
-    public void testFullRefreshForUnion()
-    {
-        String sourceTableName = "source_table" + randomNameSuffix();
-        String materializedViewName = "test_materialized_view_" + randomNameSuffix();
-
-        Session defaultSession = getSession();
-
-        String matViewDef = """
-                SELECT a, b FROM %s a WHERE a.a < 3 UNION ALL
-                SELECT * FROM %s b WHERE b.a > 5""".formatted(sourceTableName, sourceTableName);
-
-        // create source table and two identical MVs
-        assertUpdate("CREATE TABLE %s (a int, b varchar)".formatted(sourceTableName));
-        assertUpdate("INSERT INTO %s VALUES (1, 'abc'), (2, 'def')".formatted(sourceTableName), 2);
-        assertUpdate("CREATE MATERIALIZED VIEW %s AS %s".formatted(materializedViewName, matViewDef));
-
-        // execute first refresh: afterwards both MVs will contain: (1, 'abc'), (2, 'def')
-        assertUpdate("REFRESH MATERIALIZED VIEW %s".formatted(materializedViewName), 2);
-
-        // add some new rows to source
-        assertUpdate("INSERT INTO %s VALUES (3, 'ghi'), (4, 'jkl'), (5, 'mno'), (6, 'pqr')".formatted(sourceTableName), 4);
-
-        // will do a full refresh
-        assertUpdate(defaultSession, "REFRESH MATERIALIZED VIEW %s".formatted(materializedViewName), 3);
-
-        // verify that view contents are the same
-        assertThat(query("TABLE %s".formatted(materializedViewName))).matches("VALUES (1, VARCHAR 'abc'), (2, VARCHAR 'def'), (6, VARCHAR 'pqr')");
-
-        // cleanup
-        assertUpdate("DROP MATERIALIZED VIEW %s".formatted(materializedViewName));
-        assertUpdate("DROP TABLE %s".formatted(sourceTableName));
-    }
-
-    @Test
-    public void testFullRefreshForUpdates()
-    {
-        String sourceTableName = "source_table" + randomNameSuffix();
-        String materializedViewName = "test_materialized_view_" + randomNameSuffix();
-
-        Session defaultSession = getSession();
-
-        String matViewDef = "SELECT a, b FROM %s WHERE a < 3 OR a > 5".formatted(sourceTableName);
-
-        // create source table and an MV
-        assertUpdate("CREATE TABLE %s (a int, b varchar)".formatted(sourceTableName));
-        assertUpdate("INSERT INTO %s VALUES (1, 'abc'), (2, 'def')".formatted(sourceTableName), 2);
-        assertUpdate("CREATE MATERIALIZED VIEW %s AS %s".formatted(materializedViewName, matViewDef));
-
-        // execute first refresh: afterwards both MVs will contain: (1, 'abc'), (2, 'def')
-        assertUpdate("REFRESH MATERIALIZED VIEW %s".formatted(materializedViewName), 2);
-
-        // add some new rows to source
-        assertUpdate("INSERT INTO %s VALUES (3, 'ghi'), (4, 'jkl'), (5, 'mno'), (6, 'pqr')".formatted(sourceTableName), 4);
-
-        // will do incremental refresh, and only add: (6, 'pqr')
-        assertUpdate(defaultSession, "REFRESH MATERIALIZED VIEW %s".formatted(materializedViewName), 1);
-
-        // update one row and append one
-        assertUpdate("UPDATE %s SET b = 'updated' WHERE a = 1".formatted(sourceTableName), 1);
-        assertUpdate("INSERT INTO %s VALUES (7, 'stv')".formatted(sourceTableName), 1);
-
-        // will do full refresh due to the above update command
-        assertUpdate(defaultSession, "REFRESH MATERIALIZED VIEW %s".formatted(materializedViewName), 4);
-        // verify view contents
-        assertThat(query("TABLE %s".formatted(materializedViewName))).matches("VALUES (1, VARCHAR 'updated'), (2, VARCHAR 'def'), (6, VARCHAR 'pqr'), (7, VARCHAR 'stv')");
-
-        // add some new row to source
-        assertUpdate("INSERT INTO %s VALUES (8, 'wxy')".formatted(sourceTableName), 1);
-        // will do incremental refresh now since refresh window now does not contain the delete anymore, and only add: (8, 'wxy')
-        assertUpdate(defaultSession, "REFRESH MATERIALIZED VIEW %s".formatted(materializedViewName), 1);
-        // verify view contents
-        assertThat(query("TABLE %s".formatted(materializedViewName))).matches("VALUES (1, VARCHAR 'updated'), (2, VARCHAR 'def'), (6, VARCHAR 'pqr'), (7, VARCHAR 'stv'), (8, VARCHAR 'wxy')");
-
-        // cleanup
-        assertUpdate("DROP MATERIALIZED VIEW %s".formatted(materializedViewName));
-        assertUpdate("DROP TABLE %s".formatted(sourceTableName));
-    }
-
-    @Test
-    public void testRefreshWithCompaction()
-    {
-        String sourceTableName = "source_table" + randomNameSuffix();
-        String materializedViewName = "test_materialized_view_" + randomNameSuffix();
-
-        Session defaultSession = getSession();
-
-        String matViewDef = "SELECT a, b FROM %s WHERE a < 3 OR a > 5".formatted(sourceTableName);
-
-        // create source table and an MV
-        assertUpdate("CREATE TABLE %s (a int, b varchar)".formatted(sourceTableName));
-        assertUpdate("INSERT INTO %s VALUES (1, 'abc'), (2, 'def')".formatted(sourceTableName), 2);
-        assertUpdate("CREATE MATERIALIZED VIEW %s AS %s".formatted(materializedViewName, matViewDef));
-
-        // execute first refresh: afterwards both MVs will contain: (1, 'abc'), (2, 'def')
-        assertUpdate("REFRESH MATERIALIZED VIEW %s".formatted(materializedViewName), 2);
-
-        // add some new rows to source
-        assertUpdate("INSERT INTO %s VALUES (3, 'ghi'), (4, 'jkl'), (5, 'mno'), (6, 'pqr')".formatted(sourceTableName), 4);
-
-        // will do incremental refresh, and only add: (6, 'pqr')
-        assertUpdate(defaultSession, "REFRESH MATERIALIZED VIEW %s".formatted(materializedViewName), 1);
-        // verify view contents
-        assertThat(query("TABLE %s".formatted(materializedViewName))).matches("VALUES (1, VARCHAR 'abc'), (2, VARCHAR 'def'), (6, VARCHAR 'pqr')");
-
-        // run compaction - after that, refresh will update 0 rows
-        assertUpdate(defaultSession, "ALTER TABLE %s EXECUTE OPTIMIZE".formatted(sourceTableName));
-        assertUpdate(defaultSession, "REFRESH MATERIALIZED VIEW %s".formatted(materializedViewName), 0);
-        // verify view contents
-        assertThat(query("TABLE %s".formatted(materializedViewName))).matches("VALUES (1, VARCHAR 'abc'), (2, VARCHAR 'def'), (6, VARCHAR 'pqr')");
-
-        // add some new rows to source
-        assertUpdate("INSERT INTO %s VALUES (7, 'stv'), (8, 'wxy')".formatted(sourceTableName), 2);
-        // will do incremental refresh, and only add: (7, 'stv'), (8, 'wxy')
-        assertUpdate(defaultSession, "REFRESH MATERIALIZED VIEW %s".formatted(materializedViewName), 2);
-        // verify view contents
-        assertThat(query("TABLE %s".formatted(materializedViewName))).matches("VALUES (1, VARCHAR 'abc'), (2, VARCHAR 'def'), (6, VARCHAR 'pqr'), (7, VARCHAR 'stv'), (8, VARCHAR 'wxy')");
-
-        // cleanup
-        assertUpdate("DROP MATERIALIZED VIEW %s".formatted(materializedViewName));
-        assertUpdate("DROP TABLE %s".formatted(sourceTableName));
     }
 
     @Test
